refactor(trace): Extract generateRowStates into own module
Move the 133-line generateRowStates function from VirtualizedTraceView
into generateRowStates.ts, split into three focused functions:

- buildVisibleRows: main loop applying collapse + service filter pruning
- insertPrunedPlaceholders: depth-stack linear pass for placeholder rows
- generateRowStates: public entry point composing both passes

VirtualizedTraceView now imports generateRowStates and RowState from
the new module. No behavioral changes.
